actually fix cooldowns

+++ Commander-Core/src/main/java/net/octopvp/commander/annotation/Cooldown.java
@@ -10,7 +10,7 @@ import java.util.concurrent.TimeUnit;
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ElementType.METHOD, ElementType.TYPE})
 public @interface Cooldown {
-    long value();
+    double value();
 
     TimeUnit unit() default TimeUnit.SECONDS;
